+++ bindings/nostr-sdk-ffi/src/protocol/nips/mod.rs
@@ -25,6 +25,7 @@ pub mod nip56;
 pub mod nip57;
 pub mod nip59;
 pub mod nip65;
+pub mod nip73;
 pub mod nip90;
 pub mod nip94;
 pub mod nip98;