+++ crates/nostr/src/nips/mod.rs
@@ -43,6 +43,7 @@ pub mod nip58;
 #[cfg(feature = "nip59")]
 pub mod nip59;
 pub mod nip65;
+pub mod nip73;
 pub mod nip90;
 pub mod nip94;
 pub mod nip98;