+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 * nostr: add `TagStandard::Quote` ([reyamir])
 * nostr: add `Event::coordinate` ([Yuki Kishimoto])
 * nostr: add `A/a` tags in `EventBuilder::comment` (NIP22) events ([Yuki Kishimoto])
+* nostr: add NIP73 support ([Yuki Kishimoto])
 * book: add some examples ([RydalWater])
 
 ### Fixed